move all sql statements into own file
